requirements.txt
langchain
faiss-cpu
sentence-transformers
llama-cpp-python
PyMuPDF
pytesseract
Pillow
python-dotenv
langchain-community
pypdf
langchain-huggingface